+++ NodeVideoEffects.Nodes/Math/DivNode.cs
@@ -22,10 +22,10 @@ namespace NodeVideoEffects.Nodes.Math
 
         public override Task Calculate()
         {
-            if ((double)(Inputs[1].Value ?? 0) == 0 && (bool)(Inputs[2].Value ?? true))
-                Outputs[0].Value = 0.0;
+            if ((float)(Inputs[1].Value ?? 0) == 0 && (bool)(Inputs[2].Value ?? true))
+                Outputs[0].Value = 0;
             else
-                Outputs[0].Value = (double)(Inputs[0].Value ?? 0) / (double)(Inputs[1].Value ?? 0);
+                Outputs[0].Value = (float)(Inputs[0].Value ?? 0) / (float)(Inputs[1].Value ?? 0);
             return Task.CompletedTask;
         }
     }
